modules/asset-generator/requirements.txt
click>=8.1.0
pydantic>=2.0.0
Pillow>=10.0.0
Jinja2>=3.1.0
httpx>=0.25.0
anthropic>=0.40.0
pytest>=7.0.0
